net: Split public and internal protocol errors
Makes visibility more obvious.
